prismm/run_simulation/simulate_cancer_genome/tests/test_simulate_GD.py
import pytest

# Input type validation test
def test_simulate_gd_invalid_types():
    with pytest.raises(AssertionError):
        simulate_gd("invalid_type", 1, 46)

    with pytest.raises(AssertionError):
        simulate_gd({}, "invalid_type", 46)

    with pytest.raises(AssertionError):
        simulate_gd({}, 1, "invalid_type")

# Edge case with zero chromosomes
def test_simulate_gd_zero_chromosomes():
    initial_chrom_count = 0
    final_chrom_count, updated_chromosomes = simulate_gd({}, 1, initial_chrom_count)
    assert final_chrom_count == initial_chrom_count
    assert updated_chromosomes == {}

# Chromosome is dead
def test_simulate_gd_dead_chromosomes():
    chromosomes = {"type1": [{"dead": True}]}
    initial_chrom_count = 1
    final_chrom_count, updated_chromosomes = simulate_gd(chromosomes, 1, initial_chrom_count)
    assert final_chrom_count == initial_chrom_count
    assert updated_chromosomes == chromosomes

# Diverse scenarios for chromosomes
def test_simulate_gd_diverse_scenarios():
    chromosomes = {"type1": [{"dead": True}], "type2": [{"dead": False}]}
    initial_chrom_count = 2
    final_chrom_count, updated_chromosomes = simulate_gd(chromosomes, 1, initial_chrom_count)
    assert final_chrom_count == initial_chrom_count + 1  # one new chromosome should be created
